+++ src/repositories/index.ts
@@ -6,4 +6,5 @@ export { announcementsRepository } from './announcements';
 export { usersRepository } from './users';
 export { bandsRepository } from './bands';
 export { duckRepository } from './duck';
+export { reactionsRepository } from './reactions';
 export { badgeHistoryRepository } from './badgeHistoryRepository';